Prepare splitting the main build into multiple builds
There are still certain things that are done in the root of the
repository. On the long run, we should check if more can be moved
into a project responsible for a certain aspect. Then it becomes
simpler to relocate things.

We want to split up the build and also allow individual builds to work.
Therefore, there is no guarantee anymore that the repository root
is found by navigating up the project/build hierarchy. Instead, we now
search upwards for the 'version.txt' file as a marker for the root.

This can be changed and improved in the future. For now this unblocks
the transition to a composite build.

Furthermore, the global daemonTracker service is now always retrieved
from the root of the build tree.